Add functionality to show all to-dos on home view
This commit adds functionality to show all to-dos on the home view
of the project, as well as redirecting back to home page after an
item has been added to user's to-do list.

The appropriate tests for these features have also been incorporated
into 'tests.py' (unit tests) and 'functional_tests.py'.