Add a list_view to views, updates urls and tests accordingly
This commit adds a list view to the project to separate  the
home_page view from the view a user sees when accessing a to-do
list, adds an HTML template for the list view, and updates urls.py
with the url for the new list view.

At this point there is no functionality for users to have their
own to-do lists yet.